requirements.txt
streamlit
yfinance
pandas
numpy
python-dotenv
passlib[bcrypt]
google-generativeai
google-genai
